+++ nextjs/src/app/[locale]/(solutions-group)/solutions.tsx
@@ -7,6 +7,7 @@ import Intro from "@/components/intro"
 import { NavProvider } from "@/components/nav-bar/nav-context"
 import HeroWrapper from "@/components/stapi/hero-wrapper"
 import { renderSections } from "@/components/dynamic-zone/render-sections"
+import Footer from "@/components/stapi/footer"
 
 export default async function Solutions({
   activeLocale,
@@ -15,7 +16,7 @@ export default async function Solutions({
 }) {
   const url = `solutions-overview?locale=${activeLocale.locale}`
 
-  const { data } = await (await strapi(url)).json()
+  const data = await strapi(url)
   const locales = data.localizations.map((item: { locale: string }) => {
     return {
       href: `/${item.locale}/${SLUGS[item.locale]}`,
@@ -39,6 +40,7 @@ export default async function Solutions({
         </Intro>
       )}
       {sections && sections.length > 0 && sections.map(renderSections)}
+      <Footer locale={activeLocale.locale} />
     </>
   )
 }