+++ nextjs/src/app/[locale]/(solutions-group)/solutions/[slug]/page.tsx
@@ -6,6 +6,7 @@ import NavBar from "@/components/nav-bar/nav-bar"
 import { NavProvider } from "@/components/nav-bar/nav-context"
 import { renderSections } from "@/components/dynamic-zone/render-sections"
 import HeroWrapper from "@/components/stapi/hero-wrapper"
+import Footer from "@/components/stapi/footer"
 
 const SUB_PAGE = {
   en: "solutions",
@@ -24,7 +25,7 @@ export default async function SolutionsDetailPage({
     isActive: true,
   }
   const url = `solutions-pages/${slug}`
-  const { data } = await (await strapi(url)).json()
+  const data = await strapi(url)
   const locales = data.localizations.map((item: { locale: string }) => {
     return {
       href: `/${item.locale}/${SLUGS[item.locale]}`,
@@ -49,6 +50,7 @@ export default async function SolutionsDetailPage({
         </Intro>
       )}
       {sections && sections.length > 0 && sections.map(renderSections)}
+      <Footer locale={currentLocale.locale} />
     </>
   )
 }