+++ nextjs/src/app/[locale]/layout.tsx
@@ -1,11 +1,5 @@
 import { ReactNode } from "react"
-import Topbar from "@/components/topbar"
 
 export default function Layout({ children }: { children: ReactNode }) {
-  return (
-    <>
-      {/*<Topbar />*/}
-      {children}
-    </>
-  )
+  return <>{children}</>
 }