+++ nextjs/src/app/not-found.tsx
@@ -1,0 +1,12 @@
+import Container from "@/components/container"
+import React from "react"
+
+export default async function NotFound() {
+  return (
+    <Container background="neutral">
+      <div>
+        <h1>404 - Page Not Found</h1>
+      </div>
+    </Container>
+  )
+}