+++ nextjs/src/components/dynamic-zone/render-sections.tsx
@@ -12,6 +12,8 @@ import ServicesSection from "@/components/dynamic-zone/wrapper/sections/services
 import SectionSolutionsRelation from "@/components/dynamic-zone/wrapper/relations/section-solutions-relation"
 import SlaCardSection from "@/components/dynamic-zone/wrapper/relations/sla-card-section"
 import ColorCardSliderSection from "@/components/dynamic-zone/wrapper/sections/color-card-slider-section"
+import ProductCardsSection from "@/components/dynamic-zone/wrapper/relations/product-cards-section"
+import InfoDetailsSection from "@/components/dynamic-zone/wrapper/sections/info-details-section"
 
 export function renderSections(section: any, index: number) {
   switch (section.__component) {
@@ -105,6 +107,20 @@ export function renderSections(section: any, index: number) {
       return (
         <ColorCardSliderSection
           key={`sections.color-card-slider-section_${index}`}
+          section={section}
+        />
+      )
+    case "relations.product-cards-section":
+      return (
+        <ProductCardsSection
+          key={`relations.product-cards-section_${index}`}
+          section={section}
+        />
+      )
+    case "sections.info-details-section":
+      return (
+        <InfoDetailsSection
+          key={`sections.info-details-section_${index}`}
           section={section}
         />
       )