+++ nextjs/src/components/nav-bar/nav-mobile-item.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react"
+import React from "react"
 import { type NavItem } from "./nav"
 import IconChevronRight from "../icons/icon-chevron-right"
 import clsx from "clsx"
@@ -44,8 +44,8 @@ const NavMobileItem: React.FC<{
               </h3>
             </li>
 
-            {item.items &&
-              item.items.map((subItem, subIndex) => (
+            {item.menu_segment &&
+              item.menu_segment.map((subItem, subIndex) => (
                 <NavMobileSubItem key={subIndex} item={subItem} />
               ))}
           </ul>