+++ nextjs/src/components/nav-bar/nav.ts
@@ -1,5 +1,15 @@
+type Item = {
+  title: string
+  url: string
+}
+
+type Segment = {
+  title: string
+  items: Item[]
+}
+
 export type NavItem = {
   title: string
   url?: string
-  items?: NavItem[]
+  menu_segment?: Segment[]
 }