+++ nextjs/src/components/sections/section-quote.tsx
@@ -65,7 +65,7 @@ const SectionQuote: React.FC<SectionQuoteProps> = ({
                 "md:w-40 md:h-40",
                 "lg:w-60 lg:h-60",
               ])}
-              src="https://images.unsplash.com/photo-1494790108377-be9c29b29330?w=800&auto=format&fit=crop&q=60&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8M3x8YXZhdGFyfGVufDB8fDB8fHww"
+              src={image.url}
               alt={image.alt || author || ""}
               width={240}
               height={240}