+++ strapi/src/api/card-product/routes/card-product.ts
@@ -1,0 +1,7 @@
+/**
+ * card-product router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::card-product.card-product');