+++ strapi/src/api/card-product/services/card-product.ts
@@ -1,0 +1,7 @@
+/**
+ * card-product service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::card-product.card-product');