+++ strapi/src/api/navigation-menu/services/navigation-menu.ts
@@ -1,0 +1,7 @@
+/**
+ * navigation-menu service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::navigation-menu.navigation-menu');