+++ strapi/src/api/page-partner-and-product/routes/page-partner-and-product.ts
@@ -7,7 +7,9 @@ import { factories } from '@strapi/strapi';
 export default factories.createCoreRouter('api::page-partner-and-product.page-partner-and-product', {
   config: {
     findOne: {
-      middlewares: ['api::page-partner-and-product.populate-partner-and-products']
-    }
+      middlewares: [
+        'global::populate-sections-with-intro-and-localizations',
+      ],
+    },
   },
 });